Introduced credentials customization without storing them locally
without this change we acquired the username and password to Artifactory / Docker and stored it locally at `settings.xml` or `gradle.properties` files. That worked but from the security perspective wasn't the best approach. Also it was impossible to easily modify the existing settings.xml.

with this change we store all the credentials in the credentials store (Jenkins) or we take them from env vars directly (Concourse). `gradle.properties` got removed and `settings.xml` was rewritten to become parametrized. That way we can update the credentials without the need to change `settings.xml`

fixes #96